requirements.txt
torch
torchvision
matplotlib~=3.7.1
imageio~=2.28.1
tqdm~=4.65.0
numpy~=1.24.2
Pillow~=9.3.0